move lib folder out of app
