+++ lib/helpers/Encrypt.js
