+++ lib/middlewares/UserValidators.js
@@ -1,4 +1,4 @@
-const db = require('../../../database/models');
+const db = require('../../database/models');
 const reqResponses = require('../helpers/Responses');
 
 let message;
